requirements.txt
Django>=5.0
django-cities-light>=3.11
Pillow>=10.0